Move files to their normal folders
I've left only the components in the components folder and moved actions
and state to the root. We don't use the name "store" anymore to avoid
introducing an additional concept, just actions and state.
