+++ packages/frontity-org-theme/src/components/newsletter/afternewsletter.tsx
@@ -2,8 +2,8 @@ import { connect, styled } from "frontity";
 import { Connect } from "frontity/types";
 import React from "react";
 
-import FrontityOrg from "../../../../types";
-import { addAlpha } from "../../../utils";
+import FrontityOrg from "../../../types";
+import { addAlpha } from "../../utils";
 import SubmitButton from "./button";
 import Question from "./question";
