+++ packages/frontity-org-theme/src/components/newsletter/newsletter.tsx
@@ -1,9 +1,9 @@
-import { connect, css,styled } from "frontity";
+import { connect, css, styled } from "frontity";
 import { Connect } from "frontity/types";
 import React from "react";
 
-import FrontityOrg from "../../../../types";
-import { addAlpha } from "../../../utils";
+import FrontityOrg from "../../../types";
+import { addAlpha } from "../../utils";
 import SubmitButton from "./button";
 
 const Newsletter: React.FC<Connect<FrontityOrg>> = ({ state, actions }) => {
