+++ packages/frontity-org-theme/src/components/newsletter/thanks.tsx
@@ -2,7 +2,7 @@ import { connect } from "frontity";
 import { Connect } from "frontity/types";
 import React from "react";
 
-import FrontityOrg from "../../../../types";
+import FrontityOrg from "../../../types";
 
 const Thanks: React.FC<Connect<FrontityOrg, { thanksMessage: string }>> = ({
   state,
